% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{build_custom_put}
\alias{build_custom_put}
\title{Helper function to build custom S3 object readers.}
\usage{
build_custom_put(FUN, fun.defaults = NULL)
}
\value{
function
}
\description{
Helper function to build custom S3 object readers.
}
