% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_uri}
\alias{build_uri}
\title{Build an S3 uri}
\usage{
build_uri(..., dir = F)
}
\arguments{
\item{...}{components of a uri that you'd like to build}
}
\value{
character string uri
}
\description{
Internal function to quick ly generate fully formed uris from partial
uri fragments.
}
