% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.list.R
\name{dev.s3_ls}
\alias{dev.s3_ls}
\title{List objects in an S3 bucket.}
\usage{
dev.s3_ls(..., recursive = FALSE, pattern = NULL, full.names = FALSE,
  files.only = FALSE, dir.only = FALSE, all.files = FALSE,
  full.response = FALSE, aws.args = NULL)
}
\arguments{
\item{...}{flexible s3 path description of s3 location and object name. This is 
relative to your cwd (use s3_cd() to print your cwd). Accepts a character 
vector "top/next", separate character strings c("top", "next"), or 
lists list("top", "next").}

\item{recursive}{logical, when enabled lists all files and prefix directories.}

\item{pattern}{character string pattern to filter results.}

\item{full.names}{logical return fully qualified file names.}

\item{files.only}{logical filter to show only files}

\item{dir.only}{logical filter to show only directories. NOTE: Due to the fact 
that directories don't actually exist on S3, this won't work with recursive = T}

\item{all.files}{logical show unnamed objects ("")}

\item{full.response}{logical list entire metadata string from aws. Includes file
modified date/time and size.}

\item{aws.args}{character string of any additional values you need appended 
to this aws.cli command line call.}
}
\value{
character vector of bucket contents
}
\description{
List the items in an S3 bucket. Lots of time-saving filters are built-in, including
the default use of a defined working directory and grep pattern matching. As with 
all functions in this package, you must first define your environment with s3_set().
}
