% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{s3_cd}
\alias{s3_cd}
\title{Set the current working directory on S3}
\usage{
s3_cd(...)
}
\arguments{
\item{...}{flexible s3 path description of where to set the current working 
directory. Accepts a character vector "top/next", separate 
character strings c("top", "next"), or lists list("top", "next")}
}
\value{
returns the new cwd value
}
\description{
Set the current working directory on S3
}
