% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{s3_get_csv}
\alias{s3_get_csv}
\title{Read a csv table into R}
\usage{
s3_get_csv(..., fun.args = NULL, aws.args = NULL)
}
\value{
an S3 object read into R using read.table
}
\description{
Example of a custom import tool created with 
s3_get_table <- build_custom_get(FUN = read.csv)
}
