% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{s3_get_table}
\alias{s3_get_table}
\title{Read an S3 table into R}
\usage{
s3_get_table(x, ..., fun.args = NULL, aws.args = NULL)
}
\value{
an S3 object read into R using read.table
}
\description{
Read an S3 table into R
}
