% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{s3_get_with}
\alias{s3_get_with}
\title{Download an S3 object and read into R using specified write function.}
\usage{
s3_get_with(..., FUN, fun.args = NULL, aws.args = NULL)
}
\value{
S3 object
}
\description{
Download an S3 object and read into R using specified write function.
}
