% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{s3_ls}
\alias{s3_ls}
\title{List objects in an S3 bucket.}
\usage{
s3_ls(..., recursive = FALSE, pattern = NULL, full.names = FALSE,
  files.only = FALSE, dir.only = FALSE, all.files = FALSE,
  full.response = FALSE, aws.args = NULL)
}
\arguments{
\item{...}{}

\item{recursive}{logical,}

\item{pattern}{character,}

\item{files.only}{logical,}

\item{dir.only}{logical, NOTE: does not work with recursive = T}

\item{aws.args}{character,}

\item{list.names}{logical,}
}
\value{
character vector of bucket contents
}
\description{
List the items in an S3 bucket. Lots of time-saving filters are built-in, including
the default use of a defined working directory and grep pattern matching. As with 
all functions i nthis package, you must first define your environment with s3_set().
}
