% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{s3_put_table}
\alias{s3_put_table}
\title{Write an S3 table into R}
\usage{
s3_put_table(x, ..., fun.args = NULL, aws.args = NULL)
}
\value{
character string for s3 uri where an object was written with write.table
}
\description{
Write an S3 table into R
}
