% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{s3_put_with}
\alias{s3_put_with}
\title{Write an object in memory to S3 using specified write function.}
\usage{
s3_put_with(x, FUN, ..., fun.args = NULL, aws.args = NULL)
}
\value{
character string for s3 uri where an object was written
}
\description{
Write an object in memory to S3 using specified write function.
}
