% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{s3_set}
\alias{s3_set}
\title{Establish an environment which defines the current S3 bucket metadata.}
\usage{
s3_set(bucket = NULL, profile = "s3r-read-write-user", cache = NULL,
  sse = NULL, cwd = NULL, aws.args = NULL)
}
\arguments{
\item{bucket}{base name for s3 bucket e.g. "s3://bucket-name" or "bucket-name"}

\item{cwd}{character string path, optional to define a working prefix}

\item{local.cache}{character string, path of a local scratch directory. This will
be placed at /tmp/s3-cache when called the first time if none
is specified.}
}
\value{
print details about current environment
}
\description{
This environment is required for all of my S3... functions. Sets bucket name,
defines and creates a local cache directory, tracks working directory. There are
no default function arguments on purpose, to enable you to call update individual
settings without explicitly changing previous settings.
}
