% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{s3_get_save}
\alias{s3_get_save}
\title{Download an S3 object to local cache directory.}
\usage{
s3_get_save(..., aws.args = NULL)
}
\value{
file path
}
\description{
Download an S3 object to local cache directory.
}
