% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{s3_get_table}
\alias{s3_get_table}
\title{Read tables from S3 bucket directly into df object}
\usage{
s3_get_table(pre, env = e, cache = F, aws.args = "", ...)
}
\arguments{
\item{s3.path}{}
}
\value{
data.frame
}
\description{
Works for all delimited tables supported by toolboxR::Autoread().
which includes csv, tab-delim, and xlsx). Only
}
