% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_copy.R
\name{s3_mv}
\alias{s3_mv}
\title{List objects in an S3 bucket.}
\usage{
s3_mv(from = NULL, to = NULL, aws.args = NULL)
}
\arguments{
\item{aws.args}{character,}

\item{recursive}{logical,}

\item{pattern}{character,}

\item{list.names}{logical,}

\item{files.only}{logical,}

\item{dir.only}{logical,}
}
\description{
List the items in an S3 bucket. Lots of time-saving filters are built-in, including
the default use of a defined working directory and grep pattern matching. As with
all functions i nthis package, you must first define your environment with s3_set().
}
