% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{s3_put_table}
\alias{s3_put_table}
\title{Read tables from S3 bucket directly into df object}
\usage{
s3_put_table(table, pre, env = e, cache = F, aws.args = "", ...)
}
\arguments{
\item{cache}{logical value as to whether local file should be
left, default for cache=F removes file after transfer.}

\item{s3.path}{Full character string uri of the table.}
}
\value{
data.frame
}
\description{
Works for all delimited tables supported by toolboxR::auto_read().
which includes csv, tab-delim, and xlsx). Only
}
