% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_put.R
\name{s3_put_with}
\alias{s3_put_with}
\title{List objects in an S3 bucket.}
\usage{
s3_put_with(x, FUN, ..., fun.args = NULL, aws.args = NULL)
}
\arguments{
\item{x}{an object in R}

\item{FUN}{function used to write R-object to a file. Currently this is limited to 
 functions in which accept the object and file as the first argument 
(e.g. write.csv(x, file) ).}

\item{...}{character vector describing s3 location, this must end with a valid s3 object name}

\item{aws.args}{character, special arguments for writing to s3}

\item{opts}{list arguments to pass to FUN (e.g. opts = list(quote = F, row.names = F))}
}
\description{
List the items in an S3 bucket. Lots of time-saving filters are built-in, including
the default use of a defined working directory and grep pattern matching. As with
all functions i nthis package, you must first define your environment with s3_set().
}
