% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{s3_set}
\alias{s3_set}
\title{Configure s3r environment which stores S3 bucket metadata.}
\usage{
s3_set(bucket = "s3r-test-bucket", profile = "s3r-read-write-user",
  cache = NULL, sse = NULL, cwd = NULL, aws.args = NULL)
}
\arguments{
\item{bucket}{base name for s3 bucket e.g. "s3://bucket-name" or "bucket-name"}

\item{profile}{character string for the credential profile you configured in aws cli}

\item{cache}{character string for where to store local versions of downloaded s3 objects.
Defaults to current working directory.}

\item{sse}{logical should server-side encryption be used. NOTE: This does not enable
encryption on your bucket, but adds the --sse flag to write commands
 where it is required for encrypted buckets.}

\item{cwd}{character string path to define a working prefix. Defults to 
root of the bucket}

\item{aws.args}{character string of any additional values you need appended to an 
aws.cli command line call. This should likely be left blank and 
applied to individual calls where appropriate.}
}
\value{
list, invisibly returns a list of environment variable settings
}
\description{
This environment is required for all s3r functions. Sets bucket name,
defines and creates a local cache directory, tracks working directory. There are
no default function arguments on purpose, to enable you to call update individual
settings without explicitly changing previous settings.
}
